+++ client/src/__tests__/components/grid.component.test.js
@@ -2,16 +2,17 @@ import React from 'react';
 import expect from 'expect';
 import { shallow } from 'enzyme';
 import { range } from 'lodash';
-import { EMPTY_TILE } from '../constants';
+import { EMPTY_TILE } from '../../constants';
+import { getMatrixPosition } from '../../utils';
 
-import Grid from '../components/Grid';
+import Grid from '../../components/Grid';
 
 describe('Renders <Grid /> component', () => {
   const props = {
     puzzleNumbers: range(0, 16),
     shuffleNumbers: jest.fn(),
     swapNumbers: jest.fn(),
-    emptyTileIndex: EMPTY_TILE,
+    emptyTileIndex: getMatrixPosition(EMPTY_TILE),
   };
 
   const wrapper = shallow(<Grid {...props} />);
